% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D.R
\name{D}
\alias{D}
\title{This is an internal function. It calculates the criterion function for semiparametric estimation.}
\usage{
D(
  ga,
  b1,
  b2,
  b3,
  Y,
  Treat,
  R,
  u_grid = 0.02,
  e_space,
  t_tilde,
  Nonseparable = TRUE
)
}
\arguments{
\item{ga}{Parameter in the structural function.}

\item{b1}{Bandwidth for the local linear estimation of the conditional distribution of Y given T,R.}

\item{b2}{Bandwidth for the smoothing of the indicator function 1(Y <= y).}

\item{b3}{Bandwidth for the first-step conditional quantile estimation of T given R.}

\item{Y}{Outcome variable.}

\item{Treat}{Treatment variable.}

\item{R}{Running variable.}

\item{u_grid}{Grid length of the quantile space for numerical integration. The default is 0.02.}

\item{e_space}{The space of epsilon on which the numerical integration is conducted.}

\item{t_tilde}{Normalization constant in the parametric specification of the structural function.}

\item{Nonseparable}{Whether the structural function should be estimated as a separable or nonseparable function of the treatment. The default is TRUE.}
}
\value{
Returns the value of the criterion function for a given value of the parameter.
}
\description{
This is an internal function. It calculates the criterion function for semiparametric estimation.
}
