% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{This is an internal function. The structural function.}
\usage{
g(ga, t, e, t_tilde, Nonseparable = TRUE)
}
\arguments{
\item{ga}{Value of the parameter.}

\item{t}{Treatment value.}

\item{e}{Unobserved heterogeneity value.}

\item{t_tilde}{Normalization constant.}

\item{Nonseparable}{Whether the structural function should be estimated as a separable or nonseparable function of the treatment. The default is TRUE.}
}
\value{
The value of the structural function.
}
\description{
This is an internal function. The structural function.
}
