% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h0.R
\name{h0}
\alias{h0}
\title{This is an internal function. The treatment choice function below the cutoff, used for the data generating function nrddgp.}
\usage{
h0(r, u)
}
\arguments{
\item{r}{Value of the running variable.}

\item{u}{Value of the conditional rank of the treatment given the running variable.}
}
\value{
Value of the treatment choice function below the cutoff.
}
\description{
This is an internal function. The treatment choice function below the cutoff, used for the data generating function nrddgp.
}
