% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1.R
\name{h1}
\alias{h1}
\title{This is an internal function. The treatment choice function above the cutoff, used for the data generating function nrddgp.}
\usage{
h1(r, u)
}
\arguments{
\item{r}{Value of the running variable.}

\item{u}{Value of the conditional rank of the treatment given the running variable.}
}
\value{
Value of the treatment choice function above the cutoff.
}
\description{
This is an internal function. The treatment choice function above the cutoff, used for the data generating function nrddgp.
}
