% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.R
\name{k}
\alias{k}
\title{This is an internal function. The kernel function}
\usage{
k(x)
}
\arguments{
\item{x}{A numeric argument.}
}
\value{
The value of the kernel function at x.
}
\description{
This is an internal function. The kernel function
}
