% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_Y.R
\name{K_Y}
\alias{K_Y}
\title{This is an internal function. The anti-derivative of the kernel function.}
\usage{
K_Y(x)
}
\arguments{
\item{x}{A numeric argument.}
}
\value{
The value of the anti-derivative function at x.
}
\description{
This is an internal function. The anti-derivative of the kernel function.
}
