% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrdbw.R
\name{nrdbw}
\alias{nrdbw}
\title{Bandwidth selector for nonlinear RDD}
\usage{
nrdbw(df)
}
\arguments{
\item{df}{Data frame whose columns are outcome Y, treatment Treat, and running variable R, respectively.}
}
\value{
A vector of length 3 containing the recommended bandwidth for each steps in the NRDD function.
}
\description{
Provides recommendations for the three bandwidths in the semiparametric estimation of the structural function in RDD with a continuous treatment variable.
}
\examples{
n <- 500
set.seed(123)
data <- nrddgp(n)
nrdbw(data)
}
